refactor: CollectionUtils 패키지 이동
* lotto.domain -> lotto.util
