+++ src/test/java/lotto/util/CollectionUtilsTest.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.util;
 
 import static org.assertj.core.api.Assertions.*;
